Update README.md
Revealing the structure of a complex system from the observed collective dynamics is a fundamental inverse problem in network science. In the last decades, several methods have been suggested to detect the existence of interactions between the nodes of a network. By observing some physical quantities through time, the structural relationships are inferred using various discriminating statistics. With this setting, the uncertainty about the existence of the links is reflected in the uncertainty about the topological descriptors. In this study, we propose a novel methodological framework to evaluate this uncertainty, replacing the topological descriptors, even at the level of a single node, with appropriate probability distributions. Our theoretical framework agrees with the numerical experiments performed on a large set of synthetic and real-world networks. Our results provide a grounded framework for the analysis and the interpretation of widely used topological descriptors, such as degree centrality, clustering and clusters, in scenarios where the existence of network connectivity is statistically inferred.